requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
pandas
certifi
httpx
